requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
APScheduler
psycopg2-binary
alembic
pydantic
python-dotenv